[#9] feat :rocket: 로그인없이 이용하기 버튼 추가
