+++ iOS/SideDishApp/SideDishApp/Utilities/Networking/NetworkUseCase.swift
@@ -10,7 +10,7 @@ import Foundation
 
 struct NetworkUseCase {
     
-    static func makeMenu(with manager: NetworkManageable, completed: @escaping (AllMenu) -> ()) {
+    static func makeMenu(with manager: NetworkManager, completed: @escaping (AllMenu) -> ()) {
         EndPoints.allCases.map { EndPoints.BaseURL + $0.rawValue }.forEach { url in
             manager.getResource(url: url , methodType: .get, body: nil) { result in
                 switch result {
@@ -30,8 +30,7 @@ struct NetworkUseCase {
     }
     
     static func makeMenuDetail(with manager: NetworkManager, menuHash: String, completed: @escaping(MenuDetail) -> ()) {
-//        let url = EndPoints.MenuDetail + "/\(menuHash)"
-        let url = "http://www.mocky.io/v2/5ea67ef23200006700ac2a40"
+        let url = EndPoints.MenuDetail + "/\(menuHash)"
         manager.getResource(url: url, methodType: .get, body: nil) { result in
             switch result {
             case .success(let data):
