+++ iOS/SideDishApp/SideDishApp/Views/Components/MenuTableViewCell.swift
@@ -118,7 +118,7 @@ class MenuTableViewCell: UITableViewCell {
     private func makeBadgeLabel(badge: String, backgroundColor: String) -> UILabel {
         let label = UILabel()
         label.textColor = .white
-        label.backgroundColor = UIColor(hexString: backgroundColor.data(using: .utf8)!.map{ String(format: "%02x", $0)}.joined())
+        label.backgroundColor = UIColor(hexString: backgroundColor)
         label.font = UIFont.systemFont(ofSize: 14.0)
         label.text = badge
         
@@ -174,23 +174,4 @@ class MenuTableViewCell: UITableViewCell {
         
     }
     
-}
-extension UIColor {
-    convenience init(hexString: String) {
-        let hex = hexString.trimmingCharacters(in: CharacterSet.alphanumerics.inverted)
-        var int = UInt64()
-        Scanner(string: hex).scanHexInt64(&int)
-        let a, r, g, b: UInt64
-        switch hex.count {
-        case 3: // RGB (12-bit)
-            (a, r, g, b) = (255, (int >> 8) * 17, (int >> 4 & 0xF) * 17, (int & 0xF) * 17)
-        case 6: // RGB (24-bit)
-            (a, r, g, b) = (255, int >> 16, int >> 8 & 0xFF, int & 0xFF)
-        case 8: // ARGB (32-bit)
-            (a, r, g, b) = (int >> 24, int >> 16 & 0xFF, int >> 8 & 0xFF, int & 0xFF)
-        default:
-            (a, r, g, b) = (255, 0, 0, 0)
-        }
-        self.init(red: CGFloat(r) / 255, green: CGFloat(g) / 255, blue: CGFloat(b) / 255, alpha: CGFloat(a) / 255)
-    }
 }
